@@ -1,9 +1,11 @@
+import PropTypes from 'prop-types'
 import React, { Component } from 'react';
 import {
   Animated,
   Easing,
   StyleSheet,
   View,
+  ViewPropTypes,
 } from 'react-native';
 
 
@@ -63,22 +65,23 @@ class FadeInOutTransition extends Component {
 }
 
 FadeInOutTransition.defaultProps = {
-  inDuration: 250,
-  inDelay: 0,
-  outDuration: 150,
-  outDelay: 0,
   easing: Easing.linear(),
+  inDelay: 0,
+  inDuration: 250,
+  outDelay: 0,
+  outDuration: 150,
   pointerEvents: 'box-none',
+  style: null,
 };
 
 FadeInOutTransition.propTypes = {
-  style: View.propTypes.style,
-  pointerEvents: React.PropTypes.string,
-  inDuration: React.PropTypes.number,
-  inDelay: React.PropTypes.number,
-  outDuration: React.PropTypes.number,
-  outDelay: React.PropTypes.number,
-  easing: React.PropTypes.func,
+  easing: PropTypes.func,
+  inDelay: PropTypes.number,
+  inDuration: PropTypes.number,
+  outDelay: PropTypes.number,
+  outDuration: PropTypes.number,
+  pointerEvents: PropTypes.string,
+  style: ViewPropTypes.style,
 };
 
 export default FadeInOutTransition;
